refactor: fix __dirname and package.json import issues
