+++ .gitignore
@@ -14,3 +14,4 @@ redoc-static.html
 packages/cli/README.md
 .env
 packages/respect-core/src/modules/runtime-expressions/abnf-parser.cjs
+/package/
