+++ packages/core/src/benchmark/benches/lint-with-many-rules.bench.ts
@@ -7,11 +7,11 @@ import { fileURLToPath } from 'node:url';
 
 import type { StyleguideConfig } from '../../config/index.js';
 
-const __dirname = dirname(fileURLToPath(import.meta.url));
+const __internalDirname = dirname(fileURLToPath(import.meta.url ?? __dirname));
 
 export const name = 'Validate with 50 top-level rules';
 export const count = 10;
-const rebillyDefinitionRef = pathResolve(pathJoin(__dirname, 'rebilly.yaml'));
+const rebillyDefinitionRef = pathResolve(pathJoin(__internalDirname, 'rebilly.yaml'));
 const rebillyDocument = parseYamlToDocument(
   readFileSync(rebillyDefinitionRef, 'utf-8'),
   rebillyDefinitionRef
